findById 매개변수 타입 수정

+++ jjikeobang-server/src/main/java/com/jjikeobang/member/service/MemberService.java
@@ -9,7 +9,7 @@ public interface MemberService {
     List<Member> selectAllMember();
 
     void insertMember(JoinMemberDTO member);
-    Member findById(int memberId);
+    Member findById(Long memberId);
 
     Member findByLoginId(String userId);
